For Day 9 Part 2, simplify the checksum computation
Rather than reparsing the list of blocks in order to
reuse the Part 1 checksum calculation, we can calculate
checksum for each block with a bit of math.

I was hoping this would result in quicker execution,
but it doesn't appear to have a meaningful improvement
given the input size (~180ms before and after).